+++ source/board/STM32/STM32F4/STM32F4DISCOVERY/STM32F4DISCOVERY-lowLevelInitialization.cpp
@@ -26,6 +26,9 @@ namespace board
 void lowLevelInitialization()
 {
 	RCC->AHB1ENR |=
+#ifdef CONFIG_BOARD_BUTTONS_ENABLE
+			RCC_AHB1ENR_GPIOAEN |
+#endif	// def CONFIG_BOARD_LEDS_ENABLE
 #ifdef CONFIG_BOARD_LEDS_ENABLE
 			RCC_AHB1ENR_GPIODEN |
 #endif	// def CONFIG_BOARD_LEDS_ENABLE