+++ CHANGELOG.md
@@ -37,6 +37,7 @@ disabled in critical sections" (*ARCHITECTURE_ARMV7_M_KERNEL_BASEPRI*) options i
 setting, as this is a more conservative (and thus safer for a beginner) approach, which doesn't require any special
 configuration of interrupt priorities (default value is 0).
 - Support for optional inversion to all implementations of `ChipInputPin` and `ChipOutputPin`.
+- `buttons.hpp` header with board buttons for all supported boards.
 
 ### Changed